tools/hitl_testing/requirements.txt
python-can==4.0.0
pytest==7.1.1
pyyaml==6.0